fix: preserve integers within the JavaScript safe range

+++ src/cpp/web-ifc/parsing/IfcLoader.h
@@ -39,8 +39,8 @@ namespace webifc::parsing
       std::string GetDecodedStringArgument() const;
       std::string GetExpandedUUIDArgument() const;
       double GetDoubleArgument() const;
-      long GetIntArgument() const;
-      long GetIntArgument(const uint32_t tapeOffset) const;
+      int64_t GetIntArgument() const;
+      int64_t GetIntArgument(const uint32_t tapeOffset) const;
       double GetDoubleArgument(const uint32_t tapeOffset) const;
       std::string_view  GetDoubleArgumentAsString() const;
       double GetOptionalDoubleParam(double defaultValue) const;
@@ -63,7 +63,7 @@ namespace webifc::parsing
       uint32_t GetCurrentLineExpressID() const;
       void RemoveLine(const uint32_t expressID);
       void PushDouble(double input);
-      void PushInt(int input);
+      void PushInt(int64_t input);
       std::string GenerateUUID() const;
       IfcLoader* Clone();
       uint32_t GetNextExpressID(uint32_t expressId) const;
